Rama breidy

+++ src/app/home/main/main.component.ts
@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import Chart from 'chart.js/auto';
 import { Router } from '@angular/router';
 import { CasesService } from '../../services/cases.service';
+import { MaintenanceService } from '../../services/maintenance.service';
+import { UsersService } from '../../services/users.service';
+
 
 @Component({
   selector: 'app-main',
@@ -9,6 +12,7 @@ import { CasesService } from '../../services/cases.service';
   styleUrl: './main.component.css'
 })
 export class MainComponent {
+// public responseCombined: any[] = [];
 
   public token: any;
   public chart: any;
@@ -19,83 +23,149 @@ export class MainComponent {
   public responseTotals: any;
 
 
-  constructor(private _casesServ: CasesService, private _router: Router) {
+  constructor(private _casesServ: CasesService, 
+    private _router: Router, 
+    private _serMaint: MaintenanceService,
+    private _userService: UsersService
+    
+  ) {
     this.token = localStorage.getItem('token');
     // console.log(this.token);
     this.getCases();
   }
 
-  ngOnInit() {
 
-  }
+  
+
 
   getCases() {
 
     this._casesServ.casesAll(localStorage.getItem('token')).subscribe(
       response => {
         this.responseApi = response;
+        // console.log(response);
         if (this.responseApi.status == 'OK') {
-          console.log(response);
+          // console.log(response);
           this.responseCases = this.responseApi.cases;
           this.responseTotals = this.responseApi.totals;
+          // console.log(this.responseApi.totals);
           // console.log(this.responseCases);
           // console.log(this.responseTotals);
           this.generateChart();
-        } 
+        }
 
       }
 
     )
   }
 
+
+  // generateChart() {
+
+  //   // console.log(this.responseTotals);
+
+  //   let aspectRatio = 2.5;
+  //   if (window.innerWidth < 576) { // Por ejemplo, si el ancho de la ventana es menor que 576px (el punto de corte para dispositivos móviles en Bootstrap)
+  //     aspectRatio = 1.5; // Cambia el aspectRatio para dispositivos móviles
+  //   }
+
+  //   const labels = this.responseTotals.map((item: any) => item.name);
+  //   const salesData = this.responseTotals.map((item: any) => item.total_status);
+  //   // const profitData = this.responseTotals.map((item: any) => item.total_status);
+
+
+  //   const colors = [
+  //     '#ff9e18',
+  //     '#ab0a3d',
+  //     '#9e28b5',
+  //     '#0a1f8f',
+  //     '#65b2e8',
+  //     '#898b8d',
+  //     'limegreen',
+  //   ];
+
+  //   this.chart = new Chart("MyChart", {
+  //     type: 'pie',
+  //     data: {
+  //       labels: labels,
+  //       datasets: [
+  //         {
+  //           label: "Total Casos",
+  //           data: salesData,
+  //           backgroundColor: colors
+  //         },
+  //       ]
+  //     },
+  //     options: {
+  //       aspectRatio: aspectRatio,
+
+  //       //llama al evento click para redireccionar a otra ruta
+  //       onClick: (evt, activeElements) => {
+  //         if (activeElements.length > 0) {
+  //           // Obtener el índice del elemento seleccionado
+  //           const clickedIndex = activeElements[0].index;
+  //           // Obtener el nombre del elemento seleccionado
+  //           const selectedLabel = this.chart.data.labels[clickedIndex];
+  //           // Redirigir a la ruta deseada
+  //           this.redirectToRoute(selectedLabel);
+  //         }
+  //       }
+  //     },
+
+  //   });
+
+  // }
+
+
   generateChart() {
 
-    console.log(this.responseTotals);
+    let aspectRatio = 2.5;
+    if (window.innerWidth < 576) { // Por ejemplo, si el ancho de la ventana es menor que 576px (el punto de corte para dispositivos móviles en Bootstrap)
+      aspectRatio = 1.5; // Cambia el aspectRatio para dispositivos móviles
+    }
 
-    const labels = this.responseTotals.map((item: any) => item.name);
-    const salesData = this.responseTotals.map((item: any) => item.total_status);
-    // const profitData = this.responseTotals.map((item: any) => item.total_status);
+    const finalizadoData = this.responseTotals.find((item: any) => item.name === 'FINALIZADO');
 
-    this.chart = new Chart("MyChart", {
-      type: 'doughnut',
-      data: {
-        labels: labels,
-        datasets: [
-          {
-            label: "Total Casos",
-            data: salesData,
-            backgroundColor: 'red'
-          },
-          // {
-          //   label: "Profit",
-          //   data: salesData,
-          //   backgroundColor: 'limegreen'
-          // },
-          // {
-          //   label: "Profit",
-          //   data: salesData,
-          //   backgroundColor: 'pink'
-          // }
-        ]
-      },
-      options: {
-        aspectRatio: 2.5,
+      console.log(this.responseTotals);
 
-        //llama al evento click para redireccionar a otra ruta
-        onClick: (evt, activeElements) => {
-          if (activeElements.length > 0) {
-            // Obtener el índice del elemento seleccionado
-            const clickedIndex = activeElements[0].index;
-            // Obtener el nombre del elemento seleccionado
-            const selectedLabel = this.chart.data.labels[clickedIndex];
-            // Redirigir a la ruta deseada
-            this.redirectToRoute(selectedLabel);
+    if (finalizadoData) {
+      const label = finalizadoData.name;
+      const salesData = finalizadoData.total_status;
+  
+      this.chart = new Chart("MyChart", {
+        type: 'pie',
+        data: {
+          labels: [label],
+          datasets: [
+            {
+              label: "Total Casos",
+              data: [salesData],
+              backgroundColor: ['limegreen']
+            },
+          ]
+        },
+        options: {
+          aspectRatio: aspectRatio,
+  
+          //llama al evento click para redireccionar a otra ruta
+          onClick: (evt, activeElements) => {
+            if (activeElements.length > 0) {
+              // Obtener el índice del elemento seleccionado
+              const clickedIndex = activeElements[0].index;
+              // Obtener el nombre del elemento seleccionado
+              const selectedLabel = this.chart.data.labels[clickedIndex];
+              // Redirigir a la ruta deseada
+              this.redirectToRoute(selectedLabel);
+            }
           }
         }
-      }
-    });
-
+      });
+    } else {
+      console.log('No se encontraron datos para el estado "Finalizado"');
+    }
   }
+  
+
 
   // Método para redirigir a otra ruta según el elemento seleccionado
   redirectToRoute(selectedLabel: string) {
@@ -107,8 +177,11 @@ export class MainComponent {
 
 
   logout() {
-    localStorage.clear();
-    this._router.navigate(['/login']);
+    if (window.confirm('¿Está seguro de que desea salir?')) {
+      localStorage.clear();
+      this._router.navigate(['/login']);
+    }
   }
+
 
 }
